+++ src/gufo/ping/ping.py
@@ -32,6 +32,8 @@ class Ping(object):
             Use OS defaults when empty.
         coarse: Use CLOCK_MONOTONIC_COARSE when set,
             fall back to CLOCK_MONOTONIC otherwise.
+        accelerated: Enable platform-dependend accelerated
+            socket processing.
 
     Note:
         Opening the Raw Socket may require super-user priveleges
@@ -73,6 +75,7 @@ class Ping(object):
         send_buffer_size: Optional[int] = None,
         recv_buffer_size: Optional[int] = None,
         coarse: bool = False,
+        accelerated: bool = True,
     ) -> None:
         self.__size = size
         self.__ttl = ttl
@@ -81,6 +84,7 @@ class Ping(object):
         self.__send_buffer_size = send_buffer_size
         self.__recv_buffer_size = recv_buffer_size
         self.__coarse = coarse
+        self.__accelerated = accelerated
         self.__sockets: Dict[int, PingSocket] = {}
 
     @staticmethod
@@ -122,6 +126,7 @@ class Ping(object):
                 send_buffer_size=self.__send_buffer_size,
                 recv_buffer_size=self.__recv_buffer_size,
                 coarse=self.__coarse,
+                accelerated=self.__accelerated,
             )
             self.__sockets[afi] = sock
         return sock